[core/SimSettings.java] add function to retrieve information aobut sub-task
getsubTaskNum(int): get the number of subtasks of a task-based task
getsubTaskIndex(int, int): input the task type and the index of sub-task
in this task-based task, return the index of the related sub-task in
subtaskLookUpTable
getssubTaskParameter(int, int): input the task type and the index of
sub-task in this task-based task, reutrn the array of the parameters
of the subtask. This function is used in LoadGenerator.java to generate
task parameters.