[edge_client/MobileDeviceManager] submitting new sub-tasks
Since in EdgeCloudSim, when we submit a task, the task will be
scheduled. So when it comes to task-based task, the  sub-task can only
be submitted after its dependencies  have been met. So when each
sub-task finishes, the tasks that belong to the same task-based
application with it will be checked. When a task finishes in
EdgeCloudSim will execute the processCloudReturn function in
MobileDeviceManager class. Actually, CloudSim will execute the
processCloudReturn function in Broker. But the MobileDeviceManager
cannot access the variables in object LoadGeneratorModel. But the the
event of submitting new tasks can be done by sending  an event to
SimManager. In the  meantime, the sub-task need to be set as submitted.
In the class SimManager, a new label called CREATE\_SUB\_TASK is added.
And the event is handled in the function processEvent in SimManager.
Since we need to submit the new sub-tasks immediately, we use
scheduleNow function to send the event of creating new sub-tasks in the
system.